backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
pypdf==3.17.0
pinecone-client==2.2.4
openai==1.3.0
python-dotenv==1.0.0
boto3==1.28.68
# Database migration
alembic==1.11.1
# PDF processing and AI
langchain==0.0.267
faiss-cpu==1.7.4
tiktoken==0.5.1
